페이징 전환 글 조회

+++ src/main/java/com/knit/api/dto/item/ItemListResponse.java
@@ -7,5 +7,5 @@ public record ItemListResponse(
         String region,
         String status,
         String thumbnailUrl,
-        int likeCount
+        Long likeCount
 ) {}
